# Add support for YUV alpha formats