--- a/libobs/data/format_conversion.effect
+++ b/libobs/data/format_conversion.effect
@@ -31,6 +31,7 @@ uniform float3    color_range_max = {1.0, 1.0, 1.0};
 uniform texture2d image;
 uniform texture2d image1;
 uniform texture2d image2;
+uniform texture2d image3;
 
 sampler_state def_sampler {
 	Filter   = Linear;
@@ -256,6 +257,19 @@ float3 PSPlanar420_Reverse(VertTexPos frag_in) : TARGET
 	return rgb;
 }
 
+float4 PSPlanar420A_Reverse(VertTexPos frag_in) : TARGET
+{
+	int3 xy0_luma = int3(frag_in.pos.xy, 0);
+	float y = image.Load(xy0_luma).x;
+	int3 xy0_chroma = int3(frag_in.uv, 0);
+	float cb = image1.Load(xy0_chroma).x;
+	float cr = image2.Load(xy0_chroma).x;
+	float alpha = image3.Load(xy0_luma).x;
+	float3 yuv = float3(y, cb, cr);
+	float4 rgba = float4(YUV_to_RGB(yuv), alpha);
+	return rgba;
+}
+
 float3 PSPlanar422_Reverse(FragPosWide frag_in) : TARGET
 {
 	float y = image.Load(int3(frag_in.pos_wide.xz, 0)).x;
@@ -267,6 +281,19 @@ float3 PSPlanar422_Reverse(FragPosWide frag_in) : TARGET
 	return rgb;
 }
 
+float4 PSPlanar422A_Reverse(FragPosWide frag_in) : TARGET
+{
+	int3 xy0_luma = int3(frag_in.pos_wide.xz, 0);
+	float y = image.Load(xy0_luma).x;
+	int3 xy0_chroma = int3(frag_in.pos_wide.yz, 0);
+	float cb = image1.Load(xy0_chroma).x;
+	float cr = image2.Load(xy0_chroma).x;
+	float alpha = image3.Load(xy0_luma).x;
+	float3 yuv = float3(y, cb, cr);
+	float4 rgba = float4(YUV_to_RGB(yuv), alpha);
+	return rgba;
+}
+
 float3 PSPlanar444_Reverse(FragPos frag_in) : TARGET
 {
 	int3 xy0 = int3(frag_in.pos.xy, 0);
@@ -276,6 +303,25 @@ float3 PSPlanar444_Reverse(FragPos frag_in) : TARGET
 	float3 yuv = float3(y, cb, cr);
 	float3 rgb = YUV_to_RGB(yuv);
 	return rgb;
+}
+
+float4 PSPlanar444A_Reverse(FragPos frag_in) : TARGET
+{
+	int3 xy0 = int3(frag_in.pos.xy, 0);
+	float y = image.Load(xy0).x;
+	float cb = image1.Load(xy0).x;
+	float cr = image2.Load(xy0).x;
+	float alpha = image3.Load(xy0).x;
+	float3 yuv = float3(y, cb, cr);
+	float4 rgba = float4(YUV_to_RGB(yuv), alpha);
+	return rgba;
+}
+
+float4 PSAYUV_Reverse(FragPos frag_in) : TARGET
+{
+	float4 yuva = image.Load(int3(frag_in.pos.xy, 0));
+	float4 rgba = float4(YUV_to_RGB(yuva.xyz), yuva.a);
+	return rgba;
 }
 
 float3 PSNV12_Reverse(VertTexPos frag_in) : TARGET
@@ -429,6 +475,15 @@ technique I420_Reverse
 	}
 }
 
+technique I40A_Reverse
+{
+	pass
+	{
+		vertex_shader = VSTexPosHalfHalf_Reverse(id);
+		pixel_shader  = PSPlanar420A_Reverse(frag_in);
+	}
+}
+
 technique I422_Reverse
 {
 	pass
@@ -438,12 +493,39 @@ technique I422_Reverse
 	}
 }
 
+technique I42A_Reverse
+{
+	pass
+	{
+		vertex_shader = VSPosWide_Reverse(id);
+		pixel_shader  = PSPlanar422A_Reverse(frag_in);
+	}
+}
+
 technique I444_Reverse
 {
 	pass
 	{
 		vertex_shader = VSPos(id);
 		pixel_shader  = PSPlanar444_Reverse(frag_in);
+	}
+}
+
+technique YUVA_Reverse
+{
+	pass
+	{
+		vertex_shader = VSPos(id);
+		pixel_shader  = PSPlanar444A_Reverse(frag_in);
+	}
+}
+
+technique AYUV_Reverse
+{
+	pass
+	{
+		vertex_shader = VSPos(id);
+		pixel_shader  = PSAYUV_Reverse(frag_in);
 	}
 }