### Description
Enable OBS to convert YUV alpha formats to RGBA on the GPU to free up CPU.

### Motivation and Context
The GPU is much faster at processing images than the CPU.

### How Has This Been Tested?
Breakpoint inspection to verify old path (swscale MMX) has been replaced by new code paths.

Tested dancer1.webm and soccer1.webm from Google blog on Windows and Mac using Media Source. Converted those yuva420p videos to yuva422p and yuva444p using FFmpeg. AYUV tested with custom plugin since FFmpeg does not support.

Performance was measured using a custom plugin source at 1080p.

I40A -> RGBA, 1080p, CPU
swscale (MMX), Intel i7-6700: 1910 us -> 0 us

GPU measurements used Intel GPA with SetStablePowerState.

I40A -> RGBA, 1080p, GPU
Intel HD Graphics 530: 0 us -> 677 us
NVIDIA RTX 2080 Ti: 0 us -> 43 us

### Types of changes
- New feature (non-breaking change which adds functionality)
- Performance enhancement (non-breaking change which improves efficiency)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.