@jp9000 2019-08-11
Where on earth did you encounter YUVA?