@jpark37 2019-08-11
People in the Discord were talking about alpha videos and I found this: https://developers.google.com/web/updates/2013/07/Alpha-transparency-in-Chrome-video

Which points to two videos that can be downloaded here: https://simpl.info/videoalpha/

Actually, I see VLC's fourcc list calls the 4:2:0 variant "I40A", so I'll update this in a bit.