@mufunyo 2019-08-11
I do know about AYUV, which is 4:4:4:4, but other than its use in subpictures (pre-rendered subtitles) I haven't encountered it in the wild.