@jpark37 2019-08-11
Reading up a bit more, it seems like WebM never officially added alpha support. It just got hacked in, and is usable by FFmpeg and Chrome (and Adobe After Effects?).

These are from the VLC list. "I42A" and "YUVA" should be straightforward to add if desired although I haven't tried to create videos using those formats to know if they work yet.

B(VLC_CODEC_YUVA, "Planar YUV 4:4:4 Y:U:V:A"), A("YUVA"),
B(VLC_CODEC_YUV420A, "Planar YUV 4:2:0 Y:U:V:A"), A("I40A"),
B(VLC_CODEC_YUV422A, "Planar YUV 4:2:2 Y:U:V:A"), A("I42A"),