@mufunyo 2019-08-11
For reference, AYUV is documented here (it's a packed/interleaved format): https://docs.microsoft.com/en-us/windows/win32/medfound/recommended-8-bit-yuv-formats-for-video-rendering