@jpark37 2019-08-11
FFmpeg does not appear to support 8-bit AYUV judging from the header and this wiki page: https://en.wikipedia.org/wiki/FFmpeg#cite_ref-55. I wouldn't mind adding it, but I don't have a device that outputs that format; it would be custom plugin tested only.