@jp9000 2019-08-11
Oh, so this is an optimization for webms with alpha?