@jpark37 2019-08-11
It was originally when I just had yuva420p. FFmpeg doesn't want to produce WebM files with yuva422p or yuva444p, so I created some FFV1 MKVs to test those. I checked AYUV with my format testing plugin.